requirements.txt
# -----------------------------------------------------------------------------
# 招标行业 Hybrid RAG + SQL 智能问答系统 - 依赖清单
# Python 建议版本：3.9+
# -----------------------------------------------------------------------------

# Web API
fastapi>=0.115.0
uvicorn[standard]>=0.30.0

# 配置与数据校验
PyYAML>=6.0.1
pydantic>=2.7.0

# 数据库（MySQL 只读查询）
SQLAlchemy>=2.0.25
pymysql>=1.1.0

# 向量检索（本地 FAISS）
faiss-cpu>=1.8.0

# Embedding / Reranker（sentence-transformers 含 BGE 系列）
sentence-transformers>=3.0.0

# 离线预处理：PDF 文本抽取
pdfplumber>=0.11.0

# -----------------------------------------------------------------------------
# 可选：开发与测试
# -----------------------------------------------------------------------------
# pytest>=7.4.0
# pytest-cov>=4.1.0
# httpx>=0.27.0
